[#28] Control log filename with logs_filename cfg
Note that if logs_filename is used, then the deleting of old log files
will not happen. This is because the log file could be anything.
So we can't know which files 'belong' to which spider, and thus we can't
sort and delete old ones.
Perhaps this is acceptable (anyone who wants to control the log_filename
will have to manage their own log directory).