[doc update] Scrapy 0.25 breaks ScrapyD
As explained on IRC, I've had multiple feedbacks from users having issues since recently with ScrapyD, so I did a few investigations on a blank Ubuntu 14.04 machine with the official scrapinghub repository and here are the results.

Just installing with `apt-get scrapyd` results in installing the scrapy package together as a dependency, which is an alias of scrapy-0.25 and has been changed in ways that break Scrapyd (the install does not say so, but it crashes as soon as it is started saying scrapy.utils.txweb is missing).

So I tried one by one to counter-install scrapy-0.18 to scrapy-0.24 and restart scrapyd: it works in each case. scrapy-0.25 is clearly the culprit probably due to some heavy refactoring.

Therefore I'm proposing this documentation update so that users can easily understand and bypass the problem.

But a better fix would definitely be to update the ScrapyD package and make it depend on scrapy-0.24 rather than scrapy, or even better to backport the missing old code in a new version scrapy-0.26.